requirements.txt
# Core Data Processing
pyspark==3.5.0
apache-sedona==1.5.1
delta-spark==3.0.0

# Geospatial Libraries
geopandas==0.14.1
shapely==2.0.2

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Handling
pandas==2.1.3
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
